refactor: Extract stats calculation logic into clean utility module
- Create comprehensive statsUtils.ts with modular functions
- Refactor statsResolver from 200+ lines to 40 clean lines
- Separate concerns: calculateCurrentSnapshot, calculateSnapshotDifferences, etc.
- Improve code readability and maintainability for human developers
- Add proper TypeScript types and JSDoc documentation
- Make stats logic reusable and testable across the application
- Maintain all existing functionality with cleaner architecture
